Final Update - V1.0
Modified alot of things, and finally got to a point where i've done enough on this one. I have added new graphics (dont laugh lol), and a background pic which I pulled from Google Images.
I've added an options menu, and high score table etc, oh and the ability to play with 2 players and use keys or joystick.
Any comments let me know, the code got a little out of hand but its another step to learning GoDot for me..
So this is Project #3 effectively done.